refactor(ViewResolver): rename ResourceRegistry to ViewResolver
