+++ tomcat/src/main/java/nextstep/jwp/controller/GetController.java
@@ -1,30 +1,30 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.support.ResourceRegistry;
+import nextstep.jwp.servlet.ViewResolver;
 import org.apache.coyote.servlet.request.HttpRequest;
 import org.apache.coyote.servlet.response.HttpResponse;
 import org.apache.coyote.support.HttpMethod;
 
 public class GetController {
 
-    private final ResourceRegistry resourceRegistry;
+    private final ViewResolver viewResolver;
 
-    public GetController(ResourceRegistry resourceRegistry) {
-        this.resourceRegistry = resourceRegistry;
+    public GetController(ViewResolver viewResolver) {
+        this.viewResolver = viewResolver;
     }
 
     @RequestMapping(method = HttpMethod.GET, path = {"/", "/index"})
     public HttpResponse home(HttpRequest request) {
-        return resourceRegistry.findStaticResource("/index.html");
+        return viewResolver.findStaticResource("/index.html");
     }
 
     @RequestMapping(method = HttpMethod.GET, path = "/login")
     public HttpResponse login(HttpRequest request) {
-        return resourceRegistry.findStaticResource("/login.html");
+        return viewResolver.findStaticResource("/login.html");
     }
 
     @RequestMapping(method = HttpMethod.GET, path = "/register")
     public HttpResponse register(HttpRequest request) {
-        return resourceRegistry.findStaticResource("/register.html");
+        return viewResolver.findStaticResource("/register.html");
     }
 }
